+++ src/attack_flow_builder/src/assets/builder.config.ts
@@ -4,10 +4,20 @@ import { AppConfiguration } from "@/store/StoreTypes";
 import { Colors, DarkTheme } from "./scripts/BlockDiagram/DiagramFactory/Themes";
 import {
     AnchorAngle,
+    EnumPropertyDescriptor,
     PropertyType,
     SemanticRole,
     TemplateType
 } from "./scripts/BlockDiagram";
+
+const BoolEnum: EnumPropertyDescriptor = {
+    type: PropertyType.Enum,
+    options: {
+        type  : PropertyType.List,
+        form  : { type: PropertyType.String },
+        value : [["true", "True"], ["false", "False"]]
+    }
+};
 
 const config: AppConfiguration = {
     is_web_hosted: false,
@@ -23,7 +33,7 @@ const config: AppConfiguration = {
                 grid: [10, 10],
                 properties: {
                     name                         : { type: PropertyType.String, value: "Untitled Document", is_primary: true },
-                    description                  : { type: PropertyType.String, },
+                    description                  : { type: PropertyType.String },
                     author                       : {
                         type: PropertyType.Dictionary,
                         form: {
@@ -369,12 +379,7 @@ const config: AppConfiguration = {
                     name                         : { type: PropertyType.String, is_primary: true },
                     description                  : { type: PropertyType.String },
                     malware_types                : { type: PropertyType.List, form: { type: PropertyType.String, is_required: true }},
-                    is_family                    : { type: PropertyType.Enum,
-                                                     options: {
-                                                        type: PropertyType.List,
-                                                        form: { type:PropertyType.String },
-                                                        value: [["true", "True"], ["false", "False"]]},
-                                                    is_required: true },
+                    is_family                    : { ...BoolEnum, is_required: true },
                     aliases                      : { type: PropertyType.List, form: { type: PropertyType.String } },
                     kill_chain_phases            : { type: PropertyType.List, form: { type: PropertyType.String } },
                     first_seen                   : { type: PropertyType.Date },
@@ -584,12 +589,7 @@ const config: AppConfiguration = {
                 type: TemplateType.DictionaryBlock,
                 role: SemanticRole.Node,
                 properties: {
-                    is_multipart                 : { type: PropertyType.Enum,
-                        options: {
-                           type: PropertyType.List,
-                           form: { type:PropertyType.String },
-                           value: [["true", "True"], ["false", "False"]]},
-                       is_required: true },
+                    is_multipart                 : { ...BoolEnum, is_required: true },
                     date                         : { type: PropertyType.String },
                     content_type                 : { type: PropertyType.String },
                     message_id                   : { type: PropertyType.String },
@@ -673,11 +673,7 @@ const config: AppConfiguration = {
                 properties: {
                     start                        : { type: PropertyType.Date },
                     end                          : { type: PropertyType.Date },
-                    is_active                    : { type: PropertyType.Enum,
-                        options: {
-                           type: PropertyType.List,
-                           form: { type:PropertyType.String },
-                           value: [["true", "True"], ["false", "False"]]}},
+                    is_active                    : BoolEnum,
                     src_port                     : { type: PropertyType.Int, min: 0, max: 65535 },
                     dst_port                     : { type: PropertyType.Int, min: 0, max: 65535 },
                     protocols                    : { type: PropertyType.List, form: { type: PropertyType.String, is_required: true }},
@@ -696,11 +692,7 @@ const config: AppConfiguration = {
                 type: TemplateType.DictionaryBlock,
                 role: SemanticRole.Node,
                 properties: {
-                    is_hidden                    : { type: PropertyType.Enum,
-                        options: {
-                           type: PropertyType.List,
-                           form: { type:PropertyType.String },
-                           value: [["true", "True"], ["false", "False"]]}},
+                    is_hidden                    : BoolEnum,
                     pid                          : { type: PropertyType.Int, min: 0 },
                     created_time                 : { type: PropertyType.Date },
                     cwd                          : { type: PropertyType.String },
@@ -747,26 +739,10 @@ const config: AppConfiguration = {
                     account_login                : { type: PropertyType.String },
                     account_type                 : { type: PropertyType.String },
                     display_name                 : { type: PropertyType.String, is_primary: true, is_required: true },
-                    is_service_account           : { type: PropertyType.Enum,
-                                                     options: {
-                                                        type: PropertyType.List,
-                                                        form: { type:PropertyType.String },
-                                                        value: [["true", "True"], ["false", "False"]]}},
-                    is_privileged                : { type: PropertyType.Enum,
-                                                     options: {
-                                                        type: PropertyType.List,
-                                                        form: { type:PropertyType.String },
-                                                        value: [["true", "True"], ["false", "False"]]}},
-                    can_escalate_privs           : { type: PropertyType.Enum,
-                                                     options: {
-                                                        type: PropertyType.List,
-                                                        form: { type:PropertyType.String },
-                                                        value: [["true", "True"], ["false", "False"]]}},
-                    is_disabled                  : { type: PropertyType.Enum,
-                                                     options: {
-                                                        type: PropertyType.List,
-                                                        form: { type:PropertyType.String },
-                                                        value: [["true", "True"], ["false", "False"]]}},
+                    is_service_account           : BoolEnum,
+                    is_privileged                : BoolEnum,
+                    can_escalate_privs           : BoolEnum,
+                    is_disabled                  : BoolEnum,
                     account_created              : { type: PropertyType.Date },
                     account_expires              : { type: PropertyType.Date },
                     credential_last_changed      : { type: PropertyType.Date },
@@ -797,11 +773,7 @@ const config: AppConfiguration = {
                 role: SemanticRole.Node,
                 properties: {
                     subject                      : { type: PropertyType.String, is_primary: true, is_required: true },
-                    is_self_signed               : { type: PropertyType.Enum,
-                        options: {
-                           type: PropertyType.List,
-                           form: { type:PropertyType.String },
-                           value: [["true", "True"], ["false", "False"]]}},
+                    is_self_signed               : BoolEnum,
                     hashes                       : { type: PropertyType.String },
                     version                      : { type: PropertyType.String },
                     serial_number                : { type: PropertyType.String },