+++ src/attack_flow_builder/src/assets/scripts/HotkeyObserver.ts
@@ -19,11 +19,6 @@ export class HotkeyObserver<T> {
      * The function to call when a hotkey sequence is matched.
      */
     private _callback: (command: T) => void;
-    
-    /**
-     * The hotkey observer's call loop.
-     */
-    private _callLoop: number | undefined;
 
     /**
      * The DOM element the hotkey observer is watching.
@@ -39,12 +34,6 @@ export class HotkeyObserver<T> {
      * The current key state.
      */
     private _keyState: string;
-    
-    /**
-     * The amount of time (in milliseconds) a hotkey sequence needs to be held
-     * before its raised.
-     */
-    private _recognitionDelay: number;
 
 
     /**
@@ -58,11 +47,9 @@ export class HotkeyObserver<T> {
         this._boundAdvanceKeyState = this.advanceKeyState.bind(this);
         this._boundReverseKeyState = this.reverseKeyState.bind(this);
         this._callback = callback;
-        this._callLoop = undefined;
         this._container = null;
         this._hotkeyIdMap = new Map();
         this._keyState = ".";
-        this._recognitionDelay = recognitionDelay;
     }
 
 
@@ -140,27 +127,39 @@ export class HotkeyObserver<T> {
      */
     private advanceKeyState(e: KeyboardEvent) {
         let key = e.key.toLocaleLowerCase();
-        // Only acknowledge a key once
-        if (this._keyState.endsWith(`.${ key }.`))
-            return;
-        // Advanced current key state
-        clearTimeout(this._callLoop);
-        this._keyState += `${ key }.`
+        
         // If inside input field, ignore hotkeys
-        if((e.target as any).tagName === "INPUT")
+        if((e.target as any).tagName === "INPUT") {
             return;
+        }
+
+        // Advanced current key state
+        let isRepeat = this._keyState.endsWith(`.${ key }.`);
+        if(!isRepeat) {
+            this._keyState += `${ key }.`
+        }
+        
         // Check key state
         if (this._hotkeyIdMap.has(this._keyState)) {
             let hotkey = this._hotkeyIdMap.get(this._keyState)!;
-            // Disable browser default behavior if not requested
-            if (!hotkey.allowBrowserBehavior)
+            // If in repeat state and hotkey is not repeatable:
+            if(isRepeat && !hotkey.repeatable) {
+                // Prevent all browser behavior
                 e.preventDefault();
+                // Bail
+                return;    
+            }
+            // Disable browser default behavior if not requested
+            if (!hotkey.allowBrowserBehavior) {
+                e.preventDefault();
+            }
             // Execute shortcut
-            this.triggerHotkey(hotkey);
+            this._callback(hotkey.data!);
         } else {
             // If no key matched, block browser behavior by default
             e.preventDefault();
         }
+
     }
 
     /**
@@ -169,38 +168,8 @@ export class HotkeyObserver<T> {
      *  The keyup event.
      */
     private reverseKeyState(e: KeyboardEvent) {
-        clearTimeout(this._callLoop);
         let key = e.key.toLocaleLowerCase();
         this._keyState = this._keyState.replace(`.${ key }.`, ".");
-    }
-
-    /**
-     * Triggers the provided hotkey.
-     * @param hotkey
-     *  The hotkey item to trigger.
-     */
-    private triggerHotkey(hotkey: Hotkey<T>) {
-        if (!hotkey.data || hotkey.disabled)
-            return;
-        this._callLoop = setTimeout(() => {
-            // Fire hotkey
-            this._callback(hotkey.data!);
-            if (!hotkey.repeat) {
-                return;
-            }
-            // Schedule repeat call loop
-            let repeat = function (this: any) {
-                this._callback(hotkey.data!);
-                // Schedule next call
-                this._callLoop = setTimeout(
-                    repeat, hotkey.repeat!.interval
-                );
-            }.bind(this);
-            // Kick off call loop
-            this._callLoop = setTimeout(
-                repeat, hotkey.repeat!.delay
-            );
-        }, this._recognitionDelay);
     }
 
     /**
@@ -253,21 +222,9 @@ export interface Hotkey<T> {
     data?: T
 
     /**
-     * The hotkey's repeat interval.
+     * If the hotkey is repeatable.
      */
-    repeat?: {
-
-        /**
-         * The amount of time to wait (in milliseconds) before executing the
-         * interval.
-         */
-        delay: number
-
-        /**
-         * The amount of time (in milliseconds) between each successive call.
-         */
-        interval: number
-    },
+    repeatable: boolean
 
     /**
      * If the hotkey is disabled.