+++ src/attack_flow_builder/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.0.1](///compare/v2.0.0...v2.0.1) (2022-12-20)
+
+
+### Bug Fixes
+
+* add polyfill for crypto.randomUUID() 3ed71f7
+* allow copy of DateTimeField text 8cdbf01
+* enable text copying across application 5ee1ae6
+* fix issue where undo/redo menu items don't update 556c172
+* fix TextField update behavior in Firefox 1d8ee9a
+* make held hotkey behavior mirror browser's 05883b2
+* prevent tab key from focusing on disabled fields 94a054d
+
 ## 2.0.0 (2022-10-27)